+++ lib/k8s/client/runner/watch/stream.ex
@@ -5,8 +5,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
 
   alias K8s.Client.Runner.Base
   alias K8s.Client.Runner.Watch
+  alias K8s.Sys.Logger, as: K8sLogger
 
   require Logger
+
+  @log_prefix "#{__MODULE__} - " |> String.replace_leading("Elixir.", "")
 
   @type t :: %__MODULE__{
           resp: HTTPoison.AsyncResponse.t(),
@@ -77,7 +80,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
   defp next_fun({:recv, %__MODULE__{} = state}) do
     receive do
       %HTTPoison.AsyncEnd{} ->
-        Logger.warn("AsyncEnd received - tryin to restart watcher")
+        Logger.warn(
+          @log_prefix <> "AsyncEnd received - tryin to restart watcher",
+          K8sLogger.metadata()
+        )
+
         {[], {:start, state}}
 
       %HTTPoison.AsyncHeaders{} ->
@@ -89,19 +96,22 @@ defmodule K8s.Client.Runner.Watch.Stream do
         {[], {:recv, state}}
 
       %HTTPoison.AsyncStatus{code: 410} ->
-        Logger.warn("410 Gone received from watcher - trying to restart")
+        Logger.warn(
+          @log_prefix <> "410 Gone received from watcher - trying to restart",
+          K8sLogger.metadata()
+        )
+
         {[], {:start, state}}
 
       %HTTPoison.AsyncStatus{code: _} = error ->
         Logger.warn(
-          "Erronous async status received from watcher - aborting the watch",
-          error: error
+          @log_prefix <> "Erronous async status received from watcher - aborting the watch",
+          K8sLogger.metadata(error: error)
         )
 
         {:halt, nil}
 
       %HTTPoison.AsyncChunk{chunk: chunk} ->
-        Logger.debug("Chunk received")
         HTTPoison.stream_next(state.resp)
 
         {chunk, state}
@@ -109,7 +119,10 @@ defmodule K8s.Client.Runner.Watch.Stream do
         |> transform_to_events()
 
       %HTTPoison.Error{reason: {:closed, :timeout}} ->
-        Logger.debug("Request timed out - resuming the watch")
+        Logger.debug(
+          @log_prefix <> "Watch request timed out - resuming the watch",
+          K8sLogger.metadata()
+        )
 
         %{
           conn: conn,
@@ -123,7 +136,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
         {[], {:recv, %{state | resp: %HTTPoison.AsyncResponse{id: ref}}}}
 
       other ->
-        Logger.debug("Received unexpected message.", message: other)
+        Logger.debug(
+          @log_prefix <> "Wacher received unexpected message.",
+          K8sLogger.metadata(message: other)
+        )
+
         # ignore other messages and continue
         {[], {:recv, state}}
     end
@@ -139,7 +156,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
         {[], {:recv, state}}
 
       error ->
-        Logger.error("Can't restart watcher - stopping watcher stream: #{inspect(error)}")
+        Logger.error(
+          @log_prefix <> "Can't restart watcher - stopping watcher stream: #{inspect(error)}",
+          K8sLogger.metadata()
+        )
+
         {:halt, nil}
     end
   end
@@ -182,7 +203,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
         {events, {:start, state}}
 
       {:error, error}, {events, acc} ->
-        Logger.error("Could not decode JSON - chunk seems to be malformed", error: error)
+        Logger.error(
+          @log_prefix <> "Could not decode JSON - chunk seems to be malformed",
+          K8sLogger.metadata(error: error)
+        )
+
         {events, acc}
 
       {:ok, %{"object" => %{"metadata" => %{"resourceVersion" => new_resource_version}}}},
@@ -198,7 +223,11 @@ defmodule K8s.Client.Runner.Watch.Stream do
          {:recv, %__MODULE__{state | resource_version: new_resource_version}}}
 
       {:ok, %{"object" => %{"message" => message}}}, {events, {_, state}} ->
-        Logger.error("Erronous event received from watcher: #{message}")
+        Logger.error(
+          @log_prefix <> "Erronous event received from watcher: #{message}",
+          K8sLogger.metadata()
+        )
+
         {events, {:start, state}}
     end)
   end